add lvg and dependency parser to pom
update the pom file to have lvg and dependency parser in its
dependencies.
LVG and dependency parser is used for SRL feature extractor.
3 extra feature extractors are added:
PhraseExtractor -- For each basetoken, generate a string feature
indicating if this token is part of a NP or VP.
SRL Extractor -- For each basetoken, generate a string feature
indicating if it is a predicate or an argument.
WordValExtractor-- For each basetoken, generate a float feature
indicating its frquency or pca values among 3 documents of section8